+++ tests/test_helpers.py
@@ -2,6 +2,7 @@ from open_mastr.utils.helpers import (
     validate_parameter_format_for_download_method,
     validate_parameter_format_for_mastr_init,
     validate_api_credentials,
+    transform_data_parameter,
 )
 import pytest
 import sys
@@ -18,7 +19,7 @@ def db():
 def parameter_dict_working():
     parameter_dict = {
         "method": ["API", "bulk"],
-        "technology": [
+        "data": [
             "wind",
             "solar",
             "biomass",
@@ -68,20 +69,15 @@ def parameter_dict_working():
 def parameter_dict_not_working():
     parameter_dict = {
         "method": [5, "BULK", "api"],
-        "technology": [
-            "wint",
-            "Solar",
-            "biomasse",
-            5,
-        ],
+        "data": ["wint", "Solar", "biomasse", 5, []],
         "bulk_date_string": [124, "heute", 123],
         "bulk_cleansing": ["cleansing", 4, None],
         "api_processes": ["20", "None"],
         "api_limit": ["15", "None"],
         "api_date": ["None", "20220202"],
         "api_chunksize": ["20"],
-        "api_data_types": ["unite_data", 5],
-        "api_location_types": ["locatione_elec_generation", 5],
+        "api_data_types": ["unite_data", 5, []],
+        "api_location_types": ["locatione_elec_generation", 5, []],
     }
     return parameter_dict
 
@@ -97,7 +93,7 @@ def test_Mastr_validate_working_parameter(parameter_dict_working):
             parameter_dict[key] = value
             (
                 method,
-                technology,
+                data,
                 bulk_date_string,
                 bulk_cleansing,
                 api_processes,
@@ -111,7 +107,7 @@ def test_Mastr_validate_working_parameter(parameter_dict_working):
             assert (
                 validate_parameter_format_for_download_method(
                     method,
-                    technology,
+                    data,
                     bulk_date_string,
                     bulk_cleansing,
                     api_processes,
@@ -140,7 +136,7 @@ def test_Mastr_validate_not_working_parameter(
             parameter_dict[key] = value
             (
                 method,
-                technology,
+                data,
                 bulk_date_string,
                 bulk_cleansing,
                 api_processes,
@@ -153,7 +149,7 @@ def test_Mastr_validate_not_working_parameter(
             with pytest.raises(ValueError):
                 validate_parameter_format_for_download_method(
                     method,
-                    technology,
+                    data,
                     bulk_date_string,
                     bulk_cleansing,
                     api_processes,
@@ -167,7 +163,7 @@ def test_Mastr_validate_not_working_parameter(
 
 def get_parameters_from_parameter_dict(parameter_dict):
     method = parameter_dict["method"]
-    technology = parameter_dict["technology"]
+    data = parameter_dict["data"]
     bulk_date_string = parameter_dict["bulk_date_string"]
     bulk_cleansing = parameter_dict["bulk_cleansing"]
     api_processes = parameter_dict["api_processes"]
@@ -178,7 +174,7 @@ def get_parameters_from_parameter_dict(parameter_dict):
     api_location_types = parameter_dict["api_location_types"]
     return (
         method,
-        technology,
+        data,
         bulk_date_string,
         bulk_cleansing,
         api_processes,
@@ -200,6 +196,25 @@ def test_validate_parameter_format_for_mastr_init(db):
     for engine in engine_list_failing:
         with pytest.raises(ValueError):
             validate_parameter_format_for_mastr_init(engine)
+
+
+def test_transform_data_parameter(parameter_dict_working):
+    (data, api_data_types, api_location_types, harm_log,) = transform_data_parameter(
+        method="API",
+        data=["wind", "location"],
+        api_data_types=["eeg_data"],
+        api_location_types=None,
+    )
+
+    assert data == ["wind"]
+    assert api_data_types == ["eeg_data"]
+    assert api_location_types == [
+        "location_elec_generation",
+        "location_elec_consumption",
+        "location_gas_generation",
+        "location_gas_consumption",
+    ]  # TODO centralize
+    assert harm_log == ["location"]
 
 
 def test_validate_api_credentials():